web/requirements.txt
flask
pymongo
SQLAlchemy
PyMySQL
cassandra-driver==3.25.0
python-dotenv
cryptography
